Fix C043 size

+++ XAT/XAT/Game/Formats/Tmb/Entries/C043Format.cs
@@ -23,10 +23,10 @@ public class C043Format : TmbEntry
     public int Unk3 { get; set; } = 0;
 
     [UserType]
-    public int Unk4 { get; set; } = 0;
+    public short Unk4 { get; set; } = 0;
 
     [UserType]
-    public int Unk5 { get; set; } = 0;
+    public short Unk5 { get; set; } = 0;
 
     [UserType]
     public int Unk6 { get; set; } = 0;
@@ -43,8 +43,8 @@ public class C043Format : TmbEntry
         Unk1 = context.Reader.ReadInt32();
         Unk2 = context.Reader.ReadInt32();
         Unk3 = context.Reader.ReadInt32();
-        Unk4 = context.Reader.ReadInt32();
-        Unk5 = context.Reader.ReadInt32();
+        Unk4 = context.Reader.ReadInt16();
+        Unk5 = context.Reader.ReadInt16();
         Unk6 = context.Reader.ReadInt32();
     }
